+++ tools/tpcc-runner/src/tpcc.rs
@@ -23,14 +23,14 @@ pub struct RunConstants {
 #[derive(Clone, Copy, Debug)]
 pub struct TerminalAssignment {
     pub terminal_id: u32,
-    pub warehouse_id: u16,
+    pub warehouse_id: u32,
     pub district_id: u8,
 }
 
-pub fn terminal_id(warehouse_id: u16, district_id: u8) -> u32 {
+pub fn terminal_id(warehouse_id: u32, district_id: u8) -> u32 {
     debug_assert!(warehouse_id > 0);
     debug_assert!((1..=DISTRICTS_PER_WAREHOUSE).contains(&district_id));
-    (u32::from(warehouse_id) - 1) * u32::from(DISTRICTS_PER_WAREHOUSE) + u32::from(district_id)
+    (warehouse_id - 1) * u32::from(DISTRICTS_PER_WAREHOUSE) + u32::from(district_id)
 }
 
 pub fn choose_transaction<R: Rng>(rng: &mut R) -> TransactionKind {
@@ -114,23 +114,23 @@ pub fn maybe_with_original<R: Rng>(rng: &mut R, min_len: usize, max_len: usize) 
     data
 }
 
-pub fn pack_district_key(w_id: u16, d_id: u8) -> u32 {
-    (u32::from(w_id) * 100) + u32::from(d_id)
+pub fn pack_district_key(w_id: u32, d_id: u8) -> u32 {
+    (w_id * 100) + u32::from(d_id)
 }
 
-pub fn pack_customer_key(w_id: u16, d_id: u8, c_id: u32) -> u64 {
+pub fn pack_customer_key(w_id: u32, d_id: u8, c_id: u32) -> u64 {
     ((u64::from(w_id) * 100) + u64::from(d_id)) * 10_000 + u64::from(c_id)
 }
 
-pub fn pack_stock_key(w_id: u16, item_id: u32) -> u64 {
+pub fn pack_stock_key(w_id: u32, item_id: u32) -> u64 {
     u64::from(w_id) * 1_000_000 + u64::from(item_id)
 }
 
-pub fn pack_order_key(w_id: u16, d_id: u8, o_id: u32) -> u64 {
+pub fn pack_order_key(w_id: u32, d_id: u8, o_id: u32) -> u64 {
     ((u64::from(w_id) * 100) + u64::from(d_id)) * 10_000_000 + u64::from(o_id)
 }
 
-pub fn pack_order_line_key(w_id: u16, d_id: u8, o_id: u32, ol_number: u8) -> u64 {
+pub fn pack_order_line_key(w_id: u32, d_id: u8, o_id: u32, ol_number: u8) -> u64 {
     pack_order_key(w_id, d_id, o_id) * 100 + u64::from(ol_number)
 }
